Refactored TableOutputBuilder into ColumnBuilder inside TableOutput
Fixes: SIRI-120
